refactor(api): move proof to header
