+++ src/api/eth/mod.rs
@@ -1,13 +1,13 @@
 //! Ethereum API
 mod count;
 mod ffi;
-mod proof;
+mod header;
 mod proposal;
 mod receipt;
 
 pub use self::{
     count::handle as count,
-    proof::handle as proof,
+    header::handle as header,
     proposal::{codec as proposal_codec, handle as proposal, ProposalReq},
     receipt::handle as receipt,
 };
