+++ src/api/mod.rs
@@ -13,7 +13,7 @@ pub async fn serve(port: u16) -> std::io::Result<()> {
             .service(web::resource("/eth/proposal").to(eth::proposal))
             .service(web::resource("/eth/proposal/codec").to(eth::proposal_codec))
             .service(web::resource("/eth/receipt/{tx}").to(eth::receipt))
-            .service(web::resource("/eth/proof/{block}").to(eth::proof))
+            .service(web::resource("/eth/header/{block}").to(eth::header))
     })
     .disable_signals()
     .bind(format!("0.0.0.0:{}", port))?
